Refactor new allocation check logic to fit both compress and extract
