+++ client/src/hooks/datafiles/mutations/useExtract.ts
@@ -2,21 +2,9 @@ import { useMutation, useQuery } from '@tanstack/react-query';
 import { useSelector, useDispatch, shallowEqual } from 'react-redux';
 import { getExtractParams } from 'utils/getExtractParams';
 import { apiClient } from 'utils/apiClient';
-import { fetchUtil } from 'utils/fetchUtil';
 import { TTapisFile } from 'utils/types';
 import { TJobBody, TJobPostResponse } from './useSubmitJob';
-
-const getAppUtil = async function fetchAppDefinitionUtil(
-  appId: string,
-  appVersion: string
-) {
-  const params = { appId, appVersion };
-  const result = await fetchUtil({
-    url: '/api/workspace/apps',
-    params,
-  });
-  return result.response;
-};
+import { getAppUtil, getDefaultAllocation } from './toolbarAppUtils';
 
 async function submitJobUtil(body: TJobBody) {
   const res = await apiClient.post<TJobPostResponse>(
@@ -43,22 +31,12 @@ function useExtract() {
   const extractApp = useSelector(
     (state: any) => state.workbench.config.extractApp
   );
-  const defaultAllocation = useSelector((state: any) => {
-    if (state.allocations.portal_alloc) {
-      return state.allocations.portal_alloc;
-    }
-    if (
-      Array.isArray(state.allocations.active) &&
-      state.allocations.active.length > 0
-    ) {
-      return state.allocations.active[0].projectName || null;
-    }
-    return null;
-  });
-
-  const { data: latestExtract } = useQuery({
+  const { data: fullExtractApp } = useQuery({
     queryKey: ['extract-app', extractApp.id, extractApp.version],
     queryFn: () => getAppUtil(extractApp.id, extractApp.version),
+  });
+  const defaultAllocation = useSelector((state: any) => {
+    return getDefaultAllocation(state.allocations, fullExtractApp)
   });
 
   const { mutateAsync } = useMutation({ mutationFn: submitJobUtil });
@@ -77,7 +55,7 @@ function useExtract() {
     const params = getExtractParams(
       file,
       extractApp,
-      latestExtract,
+      fullExtractApp,
       defaultAllocation
     );
 
@@ -120,7 +98,7 @@ function useExtract() {
         },
         onError: (response) => {
           const errorMessage =
-            response.cause === 'compressError'
+            response.cause === 'extractError'
               ? response.message
               : 'An error has occurred.';
           dispatch({
